requirements.txt
fastapi
uvicorn
openai
python-dotenv
langserve
langchain-anthropic
langchain-google-genai
langchain-groq
